requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
python-decouple==3.8
psycopg2-binary==2.9.10
gunicorn==21.2.0
Pillow==10.0.0
django-filter==23.2
dj-database-url==2.1.0
whitenoise==6.5.0
celery==5.3.0
redis==4.6.0
